+++ line-reminder.el
@@ -284,34 +284,7 @@ If optional argument THUMBNAIL is non-nil, return in thumbnail faces."
   "Remove the indicator on LINE."
   (save-excursion
     (line-reminder--goto-line line)
-    (line-reminder--ind-remove-indicator (point))))
-
-(defun line-reminder--ind-delete-dups ()
-  "Remove duplicates for indicators overlay once."
-  (when (line-reminder--use-indicators-p)
-    (let (record-lst new-lst mkr (mkr-pos -1))
-      (dolist (ind ind-managed-absolute-indicators)
-        (setq mkr (car ind)
-              mkr-pos (marker-position mkr))
-        (if (memq mkr-pos record-lst)
-            (remove-overlays mkr-pos mkr-pos 'ind-indicator-absolute t)
-          (push mkr-pos record-lst)
-          (push ind new-lst)))
-      (setq ind-managed-absolute-indicators new-lst))))
-
-(defun line-reminder--ind-remove-indicator (pos)
-  "Remove the indicator to position POS."
-  (save-excursion
-    (goto-char pos)
-    (let ((start-pt (1+ (line-beginning-position))) (end-pt (line-end-position))
-          remove-inds)
-      (dolist (ind ind-managed-absolute-indicators)
-        (let* ((mkr (car ind)) (mkr-pos (marker-position mkr)))
-          (when (and (>= mkr-pos start-pt) (<= mkr-pos end-pt))
-            (push ind remove-inds))))
-      (dolist (ind remove-inds)
-        (setq ind-managed-absolute-indicators (remove ind ind-managed-absolute-indicators)))
-      (remove-overlays start-pt end-pt 'ind-indicator-absolute t))))
+    (remove-overlays (line-beginning-position) (line-end-position) 'ind-indicator-absolute t)))
 
 (defun line-reminder--add-change-line (line)
   "Add LINE with `modified' flag."
@@ -459,6 +432,12 @@ and END."
 (defun line-reminder--shift-all-lines (start delta)
   "Shift all `change`/`saved` lines by from START line with DELTA."
   (unless (zerop delta)
+    ;; Clean up all indicators before shifting!
+    (when (line-reminder--use-indicators-p)
+      (ht-map (lambda (line _sign)
+                (when (< start line)
+                  (line-reminder--ind-remove-indicator-at-line line)))
+              line-reminder--line-status))
     (let ((new-ht (ht-create)))
       (ht-map (lambda (line sign)
                 (if (< start line)
@@ -527,7 +506,6 @@ and END."
     ;; If buffer consider virtual buffer like `*scratch*`, then always
     ;; treat it as modified
     (setq line-reminder--undo-cancel-p (and (buffer-file-name) undo-in-progress))
-    (line-reminder--ind-delete-dups)
     (setq line-reminder--before-max-pt (point-max)
           line-reminder--before-max-linum (line-reminder--line-number-at-pos (point-max))
           line-reminder--before-begin-pt beg